+++ opush-support/src/main/java/com/lin/opush/domain/MessageTemplate.java
@@ -92,7 +92,7 @@ public class MessageTemplate implements Serializable {
     /**
      * 发送账号（邮件下可有多个发送账号、短信可有多个发送账号..）
      */
-    private Integer sendAccount;
+    private Long sendAccount;
 
     /**
      * 创建者
